Do not force update when v4.ident.me returns nothing.
Sometimes, v4.ident.me returns an empty string and the script forces
an update of the DynDNS IP which will eventually fail because no valid
IP address is specified in the request. To avoid flloding the DynDNS
service with invalid requests I added a check in one if-statement.
If the variable $ipv4 is empty then the update of the DynDNS IP will
be omitted.